+++ matchms/filtering/repair_parent_mass_from_smiles/repair_precursor_is_parent_mass.py
@@ -1,6 +1,6 @@
 import logging
 from matchms import Spectrum
-from matchms.filtering.require_parent_mass_match_smiles import _mass_diff_within_tolerance
+from matchms.filtering.repair_parent_mass_from_smiles.require_parent_mass_match_smiles import _mass_diff_within_tolerance
 from matchms.filtering.filter_utils.derive_precursor_mz_and_parent_mass import derive_precursor_mz_from_parent_mass
 
 logger = logging.getLogger("matchms")
